+++ internal/service/base/data_source_phone_delivery_settings_list.go
@@ -78,7 +78,7 @@ func (r *PhoneDeliverySettingsListDataSource) Configure(ctx context.Context, req
 		return
 	}
 
-	preparedClient, err := prepareClient(ctx, resourceConfig)
+	preparedClient, err := PrepareClient(ctx, resourceConfig)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Client not initialized",