+++ internal/service/credentials/data_source_digital_wallet_applications.go
@@ -79,7 +79,7 @@ func (r *DigitalWalletApplicationsDataSource) Configure(ctx context.Context, req
 		return
 	}
 
-	preparedClient, err := prepareClient(ctx, resourceConfig)
+	preparedClient, err := PrepareClient(ctx, resourceConfig)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Client not initialized",