+++ internal/service/mfa/resource_fido2_policy.go
@@ -502,7 +502,7 @@ func (r *FIDO2PolicyResource) Configure(ctx context.Context, req resource.Config
 		return
 	}
 
-	preparedClient, err := prepareClient(ctx, resourceConfig)
+	preparedClient, err := PrepareClient(ctx, resourceConfig)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Client not initialized",