+++ internal/service/sso/resource_population.go
@@ -109,7 +109,7 @@ func (r *PopulationResource) Configure(ctx context.Context, req resource.Configu
 		return
 	}
 
-	preparedClient, err := prepareClient(ctx, resourceConfig)
+	preparedClient, err := PrepareClient(ctx, resourceConfig)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Client not initialized",